removed validation code from badge model, moved it to Validations module, moved methods from user model to places where they needed(rules services)
