+++ app/models/user.rb
@@ -24,24 +24,6 @@ class User < ApplicationRecord
       .where(level: level, test_passages: { user_id: id })
   end
 
-  def successfully_passed_tests_by_level(level, test)
-    Test
-      .joins(:test_passages)
-      .where(level: level, test_passages: { user_id: id, correct_questions: test.questions.count })
-  end
-
-  def successfully_passed_tests_by_category(category, test)
-    Test
-      .joins(:test_passages)
-      .where(category_id: category.id, test_passages: { user_id: id, correct_questions: test.questions.count })
-  end
-
-  def successfully_passed_tests(test)
-    Test
-      .joins(:test_passages)
-      .where(test_passages: { user_id: id, correct_questions: test.questions.count })
-  end
-
   def test_passage(test)
     test_passages.order(id: :desc).find_by(test_id: test.id)
   end
